Add python dependencies

+++ rebuild.sh
@@ -19,9 +19,8 @@ if [[ $? -eq 4 ]];then
   fi
 fi
 }
-
-virsh list --name --all| egrep 'openstack-aio' | xargs -Ixx bash -c "virsh destroy xx;virsh undefine xx --nvram"
-virsh vol-list ${DISK_POOL_NAME:-data} |egrep 'openstack-aio' | egrep 'img|qcow' | awk '{print $1}' | xargs -Ixx  virsh vol-delete xx ${DISK_POOL_NAME:-data}
+test -f /usr/bin/virsh && virsh list --name --all| egrep 'openstack-aio' | xargs -Ixx bash -c "virsh destroy xx;virsh undefine xx --nvram"
+test -f /usr/bin/virsh && virsh vol-list ${DISK_POOL_NAME:-data} |egrep 'openstack-aio' | egrep 'img|qcow' | awk '{print $1}' | xargs -Ixx  virsh vol-delete xx ${DISK_POOL_NAME:-data}
 ansible-playbook -vv build_aio.yml
 echo "Sleeping 50 sec to allow boot"
 sleep 50
